Applying semifinal changes in some controllers ,services and custom exceptions

+++ src/main/java/com/lilo/service/GroupMembershipServiceImpl.java
@@ -1,7 +1,7 @@
 package com.lilo.service;
 
 import java.time.LocalDateTime;
-import java.util.List;
+import java.util.NoSuchElementException;
 
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
@@ -20,61 +20,42 @@ import com.lilo.repository.GroupMembershipRepository;
 public class GroupMembershipServiceImpl implements GroupMembershipService {
 	private GroupMembershipRepository groupMembershipRepository;
 	private KickedGroupMemberService kickedGroupMemberService;
-	private UserService userService;
 	private GroupPromotionService groupPromotionService;
 
 	public GroupMembershipServiceImpl(GroupMembershipRepository groupMembershipRepository,
-			@Lazy KickedGroupMemberService kickedGroupMemberService, @Lazy UserService userService,
+			@Lazy KickedGroupMemberService kickedGroupMemberService,
 			@Lazy GroupPromotionService groupPromotionService) {
 		this.groupMembershipRepository = groupMembershipRepository;
 		this.kickedGroupMemberService = kickedGroupMemberService;
-		this.userService = userService;
 		this.groupPromotionService = groupPromotionService;
-	}
-
-	@Override
-	public List<GroupMembership> findAllByGroupId(int groupId) {
-		List<GroupMembership> groupMemberships = groupMembershipRepository.findByGroupId(groupId);
-
-		return groupMemberships;
 	}
 
 	@Override
 	public Page<GroupMembership> findAllByGroupId(int groupId, int pageNo, int size, Sort sort) {
 		Pageable pageable = PageRequest.of(pageNo, size, sort);
-		Page<GroupMembership> page = groupMembershipRepository.findByGroupId(groupId, pageable);
-
-		return page;
+		return groupMembershipRepository.findByGroupId(groupId, pageable);
 	}
 
 	@Override
 	public GroupMembership findById(int groupMembershipId) {
-		GroupMembership groupMembership = groupMembershipRepository.findById(groupMembershipId).get();
-
-		return groupMembership;
+		return groupMembershipRepository.findById(groupMembershipId).get();
 	}
 
 	@Override
 	public GroupMembership findByUserIdAndGroupId(int userId, int groupId) {
-		GroupMembership groupMembership = groupMembershipRepository.findByUserIdAndGroupId(userId, groupId);
-
-		return groupMembership;
+		return groupMembershipRepository.findByUserIdAndGroupId(userId, groupId);
 	}
 
-//	@Override
-//	public void save(GroupMembership groupMembership) {
-//		groupMembership.setMembershipTimestamp(LocalDateTime.now());
-//		groupMembership.setRole(GroupRoles.MEMBER);
-//
-//		groupMembershipRepository.save(groupMembership);
-//	}
+	@Override
+	public boolean existsByUserIdAndGroupId(int userId, int groupId) {
+		return groupMembershipRepository.existsByUserIdAndGroupId(userId, groupId);
+	}
+
 	@Override
 	public void save(int userId, int groupId) {
 		GroupMembership membership = new GroupMembership(userId, groupId);
-
 		membership.setTimestamp(LocalDateTime.now());
 		membership.setRole(GroupRoles.MEMBER);
-
 		groupMembershipRepository.save(membership);
 	}
 
@@ -83,19 +64,12 @@ public class GroupMembershipServiceImpl implements GroupMembershipService {
 		groupMembershipRepository.save(groupMembership);
 	}
 
-//	@Override
-//	public void promote(GroupMembership groupMembership, GroupRoles groupRoles) {
-//		groupMembership.setRole(groupRoles);
-//		groupMembershipRepository.save(groupMembership);
-//
-//	}
-
 	@Override
 	public void promote(int id, int promoterId, GroupRoles newRole) {
 		GroupMembership existingMembership = groupMembershipRepository.findById(id).get();
 
 		GroupPromotion groupPromotion = new GroupPromotion(existingMembership.getGroupId(), promoterId,
-				existingMembership.getUserId(), newRole);
+				existingMembership.getUserId(), newRole, LocalDateTime.now());
 
 		existingMembership.setRole(newRole);
 		groupMembershipRepository.save(existingMembership);
@@ -103,29 +77,18 @@ public class GroupMembershipServiceImpl implements GroupMembershipService {
 	}
 
 	@Override
-	public void kick(GroupMembership groupMembership) {
-		kickedGroupMemberService.save(new KickedGroupMember(groupMembership.getUserId(), groupMembership.getGroupId()));
-
-		groupMembershipRepository.delete(groupMembership);
-	}
-
-	@Override
 	public void kick(int id) {
-		GroupMembership existingMembership = groupMembershipRepository.findById(id).get();
-		kickedGroupMemberService
-				.save(new KickedGroupMember(existingMembership.getUserId(), existingMembership.getGroupId()));
+		GroupMembership existingMembership = groupMembershipRepository.findById(id).orElseThrow();
+		kickedGroupMemberService.save(new KickedGroupMember(existingMembership.getUserId(),
+				existingMembership.getGroupId(), LocalDateTime.now()));
 		groupMembershipRepository.deleteById(id);
-	}
-
-	@Override
-	public void leave(GroupMembership groupMembership) {
-		groupMembershipRepository.delete(groupMembership);
 	}
 
 	@Override
 	public void leave(int id) {
+		if (!groupMembershipRepository.existsById(id))
+			throw new NoSuchElementException("cannot delete a non existing group membership.");
 		groupMembershipRepository.deleteById(id);
-		;
 	}
 
 }
